feat: add test

+++ src/shared/services/aws/delete.s3.ts
@@ -2,6 +2,7 @@ import { s3 } from "./config";
 
 export default {
   delete: (name_file: string) => {
-      s3.deleteObject({ Bucket: process.env.BUCKET_NAME, Key: name_file });
-  }
-}
+    if (process.env.NODE_ENV !== "production") return;
+    s3.deleteObject({ Bucket: process.env.BUCKET_NAME, Key: name_file });
+  },
+};
